fixed card validation
